validate*: do not validate output of a failed command
Although all these scripts have the `#!/bin/sh -ef` shebang, which means
they have the `set -e` (= `set -o errexit`) option set, the exit code of
a command whose output is directly used as an argument to another
program is apparently not checked. It helps to first save the output to
a variable on a separate line. Now, if the command to obtain the version
of a particular program fails, the script terminates and the
`./validate-version` script is not run.